refactor[#68] : TagSelectToggle view 세그먼트로 이동
