+++ blog/src/views/write/lib/index.ts
@@ -1,1 +1,2 @@
 export * from "./useSeriesSelectToggle";
+export * from "./useTagSelectToggle";
