--- blog/src/widgets/tag/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./TagSelectToggle";
